Setting up
Because I am to stupid to just use the environment file on windows, because it contains stuff, that is not available for windows and I cannot install linux, I am getting the necessary packages 1 by 1. Not a great idea, but so far it seems to be working. Next I will have to change the LLM used for image annotation. Man this takes ages.